api/requirements.txt
# Azure Functions
azure-functions>=1.17.0

# LangChain Core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.5

# Azure Services
azure-cosmos>=4.5.1
azure-identity>=1.15.0

# Data Validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
